Extract generic KeyValue storage from FlatFile
